# apps/server/openapi/public-v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/words": {
      "get": {
        "description": "Every enum filter accepts a repeated key; values of one filter are OR-ed, filters are AND-ed. `search` keeps the headwords starting with the prefix (an autocomplete or an A–Z browser: cacheable, unlike the search). Pass `meta.next_cursor` back as `cursor` to read the next page.",
        "operationId": "PublicWordsController_list",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Headword prefix, case-insensitive (`ru` lists run, rung, runner, …)",
            "schema": {
              "maxLength": 128,
              "type": "string"
            }
          },
          {
            "name": "is_obsolete",
            "required": false,
            "in": "query",
            "description": "true: obsolete entries only, false: current ones only",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "part_of_speech",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "noun",
                  "verb",
                  "modal_verb",
                  "adjective",
                  "adverb",
                  "pronoun",
                  "numeral",
                  "numeral_fractional",
                  "determiner",
                  "interjection",
                  "article",
                  "preposition",
                  "conjunction",
                  "letter",
                  "phrase",
                  "grammar_pattern"
                ]
              }
            }
          },
          {
            "name": "word_level",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "A1",
                  "A2",
                  "B1",
                  "B2",
                  "C1",
                  "C2"
                ]
              }
            }
          },
          {
            "name": "language_register",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "formal",
                  "informal",
                  "slang"
                ]
              }
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Words tagged with any of the categories",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "scientific",
                  "technical",
                  "medical",
                  "legal",
                  "business",
                  "IT",
                  "art",
                  "political",
                  "sport",
                  "culinary"
                ]
              }
            }
          },
          {
            "name": "area_variant",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "common",
                  "british",
                  "american",
                  "australian"
                ]
              }
            }
          },
          {
            "name": "form_of_word",
            "required": false,
            "in": "query",
            "description": "Defaults to base forms only",
            "schema": {
              "default": [
                "base_form"
              ],
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "base_form",
                  "plural_form",
                  "possessive_singular_form",
                  "possessive_plural_form",
                  "past_simple",
                  "past_participle",
                  "present_participle",
                  "third_person_singular",
                  "comparative_form",
                  "superlative_form",
                  "object",
                  "possessive_adjective",
                  "possessive_pronoun",
                  "reflexive",
                  "ordinal",
                  "multiplicative"
                ]
              }
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "The `meta.next_cursor` of the previous page; omit for the first page",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "with_meanings",
            "required": false,
            "in": "query",
            "description": "Join the meanings of every word",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "name": "with_translations",
            "required": false,
            "in": "query",
            "description": "Join the short translations of every word",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicWordsV1ResT"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input: an unknown field, a value outside the allowed set, or a foreign cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit of the public prefix exceeded (PUBLIC_API_RATE_LIMIT); retry after the window",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          }
        },
        "summary": "List dictionary entries by filters, cursor-paged and ordered by (word, id)",
        "tags": [
          "Public API v1"
        ]
      }
    },
    "/api/v1/words/batch": {
      "post": {
        "description": "Every spelling is matched like `GET /words/{word}`; the answer keeps the request order, collapses duplicates, and lists the spellings without an entry under `meta.not_found` instead of failing. One request against the rate limit, whatever the size of the batch.",
        "operationId": "PublicWordsController_batch",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WordsBatchV1ReqDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicWordsBatchV1ResT"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input: an unknown field, a value outside the allowed set, or a foreign cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit of the public prefix exceeded (PUBLIC_API_RATE_LIMIT); retry after the window",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          }
        },
        "summary": "Up to 50 headwords in one request",
        "tags": [
          "Public API v1"
        ]
      }
    },
    "/api/v1/words/id/{id}": {
      "get": {
        "operationId": "PublicWordsController_byId",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicWordV1ResT"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input: an unknown field, a value outside the allowed set, or a foreign cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          },
          "404": {
            "description": "Nothing matches",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit of the public prefix exceeded (PUBLIC_API_RATE_LIMIT); retry after the window",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          }
        },
        "summary": "One dictionary entry by its numeric id",
        "tags": [
          "Public API v1"
        ]
      }
    },
    "/api/v1/words/{word}": {
      "get": {
        "operationId": "PublicWordsController_byHeadword",
        "parameters": [
          {
            "name": "word",
            "required": true,
            "in": "path",
            "description": "Headword spelling, case-insensitive (spaces URL-encoded for phrases). An inflected form resolves to its base entry",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicHeadwordV1ResT"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input: an unknown field, a value outside the allowed set, or a foreign cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          },
          "404": {
            "description": "Nothing matches",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit of the public prefix exceeded (PUBLIC_API_RATE_LIMIT); retry after the window",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          }
        },
        "summary": "All entries of a headword: parts of speech, forms, meanings, translations, links",
        "tags": [
          "Public API v1"
        ]
      }
    },
    "/api/v1/words/{word}/meanings": {
      "get": {
        "operationId": "PublicWordsController_meanings",
        "parameters": [
          {
            "name": "word",
            "required": true,
            "in": "path",
            "description": "Headword spelling, case-insensitive (spaces URL-encoded for phrases). An inflected form resolves to its base entry",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicHeadwordMeaningsV1ResT"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input: an unknown field, a value outside the allowed set, or a foreign cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          },
          "404": {
            "description": "Nothing matches",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit of the public prefix exceeded (PUBLIC_API_RATE_LIMIT); retry after the window",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          }
        },
        "summary": "The meanings of a headword across its entries",
        "tags": [
          "Public API v1"
        ]
      }
    },
    "/api/v1/words/{word}/synonyms": {
      "get": {
        "description": "One item per linked headword and meaning; each names its `meaning_id`, `word_id` and `part_of_speech`.",
        "operationId": "PublicWordsController_synonyms",
        "parameters": [
          {
            "name": "word",
            "required": true,
            "in": "path",
            "description": "Headword spelling, case-insensitive (spaces URL-encoded for phrases). An inflected form resolves to its base entry",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicHeadwordLinksV1ResT"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input: an unknown field, a value outside the allowed set, or a foreign cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          },
          "404": {
            "description": "Nothing matches",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit of the public prefix exceeded (PUBLIC_API_RATE_LIMIT); retry after the window",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          }
        },
        "summary": "The synonyms of a headword, per meaning",
        "tags": [
          "Public API v1"
        ]
      }
    },
    "/api/v1/words/{word}/antonyms": {
      "get": {
        "description": "One item per linked headword and meaning; each names its `meaning_id`, `word_id` and `part_of_speech`.",
        "operationId": "PublicWordsController_antonyms",
        "parameters": [
          {
            "name": "word",
            "required": true,
            "in": "path",
            "description": "Headword spelling, case-insensitive (spaces URL-encoded for phrases). An inflected form resolves to its base entry",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicHeadwordLinksV1ResT"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input: an unknown field, a value outside the allowed set, or a foreign cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          },
          "404": {
            "description": "Nothing matches",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit of the public prefix exceeded (PUBLIC_API_RATE_LIMIT); retry after the window",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          }
        },
        "summary": "The antonyms of a headword, per meaning",
        "tags": [
          "Public API v1"
        ]
      }
    },
    "/api/v1/words/{word}/translations": {
      "get": {
        "operationId": "PublicWordsController_translations",
        "parameters": [
          {
            "name": "word",
            "required": true,
            "in": "path",
            "description": "Headword spelling, case-insensitive (spaces URL-encoded for phrases). An inflected form resolves to its base entry",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          },
          {
            "name": "language",
            "required": false,
            "in": "query",
            "description": "Translation languages; no value means all of them",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "ru",
                  "es",
                  "fr",
                  "de",
                  "pt",
                  "zh",
                  "ar"
                ]
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicHeadwordTranslationsV1ResT"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input: an unknown field, a value outside the allowed set, or a foreign cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          },
          "404": {
            "description": "Nothing matches",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit of the public prefix exceeded (PUBLIC_API_RATE_LIMIT); retry after the window",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          }
        },
        "summary": "The short and per-meaning translations of a headword",
        "tags": [
          "Public API v1"
        ]
      }
    },
    "/api/v1/words/{word}/history": {
      "get": {
        "description": "The edits that still show in what is served, the latest first, with the values before and after: the entries of such a headword carry `modified: true`. Empty for a headword served as its source has it.",
        "operationId": "PublicWordsController_history",
        "parameters": [
          {
            "name": "word",
            "required": true,
            "in": "path",
            "description": "Headword spelling, case-insensitive (spaces URL-encoded for phrases). An inflected form resolves to its base entry",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicHeadwordHistoryV1ResT"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input: an unknown field, a value outside the allowed set, or a foreign cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          },
          "404": {
            "description": "Nothing matches",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit of the public prefix exceeded (PUBLIC_API_RATE_LIMIT); retry after the window",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          }
        },
        "summary": "What was changed or added on the instance in the entries of a headword",
        "tags": [
          "Public API v1"
        ]
      }
    },
    "/api/v1/words/{word}/forms": {
      "get": {
        "operationId": "PublicWordsController_forms",
        "parameters": [
          {
            "name": "word",
            "required": true,
            "in": "path",
            "description": "Headword spelling, case-insensitive (spaces URL-encoded for phrases). An inflected form resolves to its base entry",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicHeadwordFormsV1ResT"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input: an unknown field, a value outside the allowed set, or a foreign cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          },
          "404": {
            "description": "Nothing matches",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit of the public prefix exceeded (PUBLIC_API_RATE_LIMIT); retry after the window",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicApiErrorT"
                }
              }
            }
          }
        },
        "summary": "The inflected forms of a headword across its entries",
        "tags": [
          "Public API v1"
        ]
      }
    },
...
  },
  "info": {
    "title": "VocabBloom Public API",
    "description": "Read-only dictionary API under /api/v1 (contract version 1): no authentication, every answer in a { data, meta } envelope, errors as { statusCode, message, error: true }, X-API-Version on every response. Successful GET answers carry ETag, Last-Modified and Cache-Control. Documentation: https://vocab-bloom-hub.com/en/docs/api.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "WordsBatchV1ReqDTO": {
        "type": "object",
        "properties": {
          "words": {
            "minItems": 1,
            "maxItems": 50,
            "description": "Headword spellings, 50 at most, each matched like GET /words/{word} (case-insensitively; an inflected form resolves to its base entry)",
            "example": [
              "run",
              "ran",
              "put up with"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "words"
        ]
      },
...
      "PublicWordV1ResT": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PublicWordV1T"
          }
        },
        "required": [
          "data"
        ]
      },
...
      "PublicHeadwordV1ResT": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicWordV1T"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PublicHeadwordV1MetaT"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "PublicHeadwordMeaningsV1ResT": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicMeaningV1T"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PublicHeadwordV1MetaT"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "PublicHeadwordLinksV1ResT": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicWordLinkV1T"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PublicHeadwordV1MetaT"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "PublicHeadwordTranslationsV1ResT": {
        "type": "object",
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/PublicHeadwordV1MetaT"
          },
          "data": {
            "$ref": "#/components/schemas/PublicHeadwordTranslationsV1T"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "PublicWordsBatchV1ResT": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicWordsBatchItemV1T"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PublicWordsBatchV1MetaT"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "PublicWordsV1ResT": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicWordV1T"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PublicWordsV1MetaT"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "PublicApiErrorT": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number"
          },
          "message": {
            "type": "string"
          },
          "error": {
            "type": "boolean"
          }
        },
        "required": [
          "error",
          "message",
          "statusCode"
        ]
      }
    }
  },
...
}